back_end/requirements.txt
fastapi
uvicorn
torch
ultralytics
numpy
opencv-python-headless
minio
psycopg2-binary
mlflow
python-multipart